+++ Assets/Scripts/Gameplay/Action/BaseTypes/ActionBase.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace Unity.Multiplayer.Samples.BossRoom
+namespace Unity.Multiplayer.Samples.BossRoom.Actions
 {
     /// <summary>
     /// Abstract base class containing some common members shared by Action (server) and ActionFX (client visual)
