+++ Assets/Scripts/Gameplay/Action/ConcreteActions/AoeActionFX.cs
@@ -1,7 +1,8 @@
 using System;
+using Unity.Multiplayer.Samples.BossRoom.Visual;
 using UnityEngine;
 
-namespace Unity.Multiplayer.Samples.BossRoom.Visual
+namespace Unity.Multiplayer.Samples.BossRoom.Actions
 {
     /// Final step in the AoE action flow. Please see AoEActionInput for the first step and more details on overall flow
     public class AoeActionFX : ActionFX
@@ -9,14 +10,14 @@ namespace Unity.Multiplayer.Samples.BossRoom.Visual
         public AoeActionFX(ref ActionRequestData data, ClientCharacterVisualization parent)
             : base(ref data, parent) { }
 
-        public override bool Start()
+        public override bool OnStart()
         {
-            base.Start();
+            base.OnStart();
             GameObject.Instantiate(Description.Spawns[0], m_Data.Position, Quaternion.identity);
             return ActionConclusion.Stop;
         }
 
-        public override bool Update()
+        public override bool OnUpdate()
         {
             throw new Exception("This should not execute");
         }
