+++ Assets/Scripts/Gameplay/Action/ConcreteActions/EmoteAction.cs
@@ -1,6 +1,7 @@
 using System;
+using Unity.Multiplayer.Samples.BossRoom.Server;
 
-namespace Unity.Multiplayer.Samples.BossRoom.Server
+namespace Unity.Multiplayer.Samples.BossRoom.Actions
 {
     public class EmoteAction : Action
     {
@@ -8,13 +9,13 @@ namespace Unity.Multiplayer.Samples.BossRoom.Server
         {
         }
 
-        public override bool Start()
+        public override bool OnStart()
         {
             m_Parent.serverAnimationHandler.NetworkAnimator.SetTrigger(Description.Anim);
             return false;
         }
 
-        public override bool Update()
+        public override bool OnUpdate()
         {
             // since we return false at Start(), this method should not execute
             throw new InvalidOperationException("No logic defined.");
