+++ Assets/Scripts/Gameplay/Action/ConcreteActions/TossAction.cs
@@ -1,8 +1,9 @@
 using UnityEngine;
 using BossRoom.Scripts.Shared.Net.NetworkObjectPool;
+using Unity.Multiplayer.Samples.BossRoom.Server;
 using Unity.Netcode;
 
-namespace Unity.Multiplayer.Samples.BossRoom.Server
+namespace Unity.Multiplayer.Samples.BossRoom.Actions
 {
     /// <summary>
     /// Action responsible for creating a physics-based thrown object.
@@ -13,7 +14,7 @@ namespace Unity.Multiplayer.Samples.BossRoom.Server
 
         public TossAction(ServerCharacter parent, ref ActionRequestData data) : base(parent, ref data) { }
 
-        public override bool Start()
+        public override bool OnStart()
         {
             // snap to face the direction we're firing
 
@@ -42,7 +43,7 @@ namespace Unity.Multiplayer.Samples.BossRoom.Server
             return true;
         }
 
-        public override bool Update()
+        public override bool OnUpdate()
         {
             if (TimeRunning >= Description.ExecTimeSeconds && !m_Launched)
             {
