+++ Assets/Scripts/Gameplay/Configuration/ActionDescription.cs
@@ -1,4 +1,5 @@
 using System;
+using Unity.Multiplayer.Samples.BossRoom.Actions;
 using UnityEngine;
 
 namespace Unity.Multiplayer.Samples.BossRoom
