+++ Assets/Scripts/Gameplay/GameplayObjects/Character/AI/AIBrain.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Unity.Multiplayer.Samples.BossRoom.Actions;
 using UnityEngine;
 
 namespace Unity.Multiplayer.Samples.BossRoom.Server
