+++ Assets/Scripts/Gameplay/GameplayObjects/Character/AI/AttackAIState.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using Unity.Multiplayer.Samples.BossRoom.Actions;
 using UnityEngine;
 
 namespace Unity.Multiplayer.Samples.BossRoom.Server
