+++ Assets/Scripts/Gameplay/UI/HeroEmoteBar.cs
@@ -1,3 +1,4 @@
+using Unity.Multiplayer.Samples.BossRoom.Actions;
 using Unity.Multiplayer.Samples.BossRoom.Client;
 using UnityEngine;
 using SkillTriggerStyle = Unity.Multiplayer.Samples.BossRoom.Client.ClientInputSender.SkillTriggerStyle;
