+++ Assets/Scripts/Gameplay/UI/UICharSelectClassInfoBox.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 using UnityEngine.UI;
 using TMPro;
-
+using Unity.Multiplayer.Samples.BossRoom.Actions;
 
 namespace Unity.Multiplayer.Samples.BossRoom.Client
 {
